+++ pkg/task/service/service.go
@@ -13,22 +13,22 @@ import (
 	"github.com/flant/addon-operator/pkg/kube_config_manager"
 	"github.com/flant/addon-operator/pkg/module_manager"
 	"github.com/flant/addon-operator/pkg/task"
-	applykubeconfigvalues "github.com/flant/addon-operator/pkg/task/apply-kube-config-values"
-	convergemodules "github.com/flant/addon-operator/pkg/task/converge-modules"
 	discovercrds "github.com/flant/addon-operator/pkg/task/discover-crds"
-	discoverhelmrelease "github.com/flant/addon-operator/pkg/task/discover-helm-release"
-	globalhookenablekubernetesbindings "github.com/flant/addon-operator/pkg/task/global-hook-enable-kubernetes-bindings"
-	globalhookenableschedulebindings "github.com/flant/addon-operator/pkg/task/global-hook-enable-schedule-bindings"
-	globalhookrun "github.com/flant/addon-operator/pkg/task/global-hook-run"
-	globalhookwaitkubernetessynchronization "github.com/flant/addon-operator/pkg/task/global-hook-wait-kubernetes-synchronization"
-	moduledelete "github.com/flant/addon-operator/pkg/task/module-delete"
-	moduleensurecrds "github.com/flant/addon-operator/pkg/task/module-ensure-crds"
-	modulehookrun "github.com/flant/addon-operator/pkg/task/module-hook-run"
-	modulepurge "github.com/flant/addon-operator/pkg/task/module-purge"
-	modulerun "github.com/flant/addon-operator/pkg/task/module-run"
 	paralleltask "github.com/flant/addon-operator/pkg/task/parallel"
-	parallelmodulerun "github.com/flant/addon-operator/pkg/task/parallel-module-run"
 	taskqueue "github.com/flant/addon-operator/pkg/task/queue"
+	applykubeconfigvalues "github.com/flant/addon-operator/pkg/task/tasks/apply-kube-config-values"
+	convergemodules "github.com/flant/addon-operator/pkg/task/tasks/converge-modules"
+	discoverhelmrelease "github.com/flant/addon-operator/pkg/task/tasks/discover-helm-release"
+	globalhookenablekubernetesbindings "github.com/flant/addon-operator/pkg/task/tasks/global-hook-enable-kubernetes-bindings"
+	globalhookenableschedulebindings "github.com/flant/addon-operator/pkg/task/tasks/global-hook-enable-schedule-bindings"
+	globalhookrun "github.com/flant/addon-operator/pkg/task/tasks/global-hook-run"
+	globalhookwaitkubernetessynchronization "github.com/flant/addon-operator/pkg/task/tasks/global-hook-wait-kubernetes-synchronization"
+	moduledelete "github.com/flant/addon-operator/pkg/task/tasks/module-delete"
+	moduleensurecrds "github.com/flant/addon-operator/pkg/task/tasks/module-ensure-crds"
+	modulehookrun "github.com/flant/addon-operator/pkg/task/tasks/module-hook-run"
+	modulepurge "github.com/flant/addon-operator/pkg/task/tasks/module-purge"
+	modulerun "github.com/flant/addon-operator/pkg/task/tasks/module-run"
+	parallelmodulerun "github.com/flant/addon-operator/pkg/task/tasks/parallel-module-run"
 	"github.com/flant/addon-operator/pkg/utils"
 	klient "github.com/flant/kube-client/client"
 	"github.com/flant/shell-operator/pkg/metric"
