+++ pkg/task/tasks/global-hook-enable-kubernetes-bindings/task.go
