+++ pkg/task/tasks/global-hook-enable-schedule-bindings/task.go
