+++ pkg/task/tasks/global-hook-wait-kubernetes-synchronization/task.go
