+++ pkg/task/tasks/module-delete/task.go
